requirements.txt
aiohttp==3.13.3
pillow==12.1.1
opencv-python==4.13.0.92
